Complete run through of placement chapter without TODOs!
Another re-run of the scalability experiments are currently running and results
are expected to fill in the missing results for million-core problems (and some
of the smaller, badly placed problems too). We shall see how they go tomorrow.
Failing that I can pull in previously gathered results and leave the routing
table results blank for the larger experiments.